more css added.
And this is a longer commit message. I usually try to keep this to
less than 78 characters across, just as a basic unix convention. In this
case, it'll also render well on github.com for the commit messages too

We could also close out issues and cross reference other things here and
a good example of this could be showing where a bug may have started.